requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
python-dotenv
httpx
requests
apscheduler
lxml
greenlet
aiosqlite
